+++ src/main/kotlin/domain/layout/LayoutWriter.kt
@@ -1,6 +1,6 @@
 package domain.layout
 
-import domain.file.Page
+import domain.model.Page
 import kotlinx.coroutines.Dispatchers.IO
 import kotlinx.coroutines.withContext
 import java.io.RandomAccessFile
