+++ src/main/kotlin/domain/model/Table.kt
@@ -1,7 +1,5 @@
 package domain.model
 
-import domain.file.Page
-
 data class Table(
     val name: String = "",
     val data: List<Page> = ArrayList()
